src/test/java/com/crm/qa/testcases/HomePageTest.java
package com.crm.qa.testcases;

import java.io.IOException;

import org.junit.Assert;
import org.testng.annotations.AfterMethod;
import org.testng.annotations.BeforeMethod;
import org.testng.annotations.Test;

import com.crm.qa.base.TestBase;
import com.crm.qa.pages.ContactsPage;
import com.crm.qa.pages.HomePage;
import com.crm.qa.pages.LoginPage;
import com.crm.qa.util.TestUtil;

public class HomePageTest extends TestBase {

	LoginPage loginPage;
	HomePage homePage;
	TestUtil testUtil;
	ContactsPage contactsPage;

	public HomePageTest() throws IOException {
		super();
	}

	@BeforeMethod
	public void setUp() throws IOException {
		initialization();
		loginPage = new LoginPage();
		testUtil = new TestUtil();
		contactsPage = new ContactsPage();
		homePage = loginPage.Login(prop.getProperty("username"), prop.getProperty("password"));
	}

	@Test(priority = 1)
	public void verifyHomePageTitleTest() {
		String homePageTitle = homePage.verifyHomePageTitle();
		Assert.assertEquals(homePageTitle, "CRMPRO");
	}

	@Test(priority =2)
	public void verifyUsernameTest() {
		testUtil.switchToFrame();
		boolean status = homePage.verifyUsernameLabel();
		Assert.assertTrue(status);
	}
	
	@Test
	public void verifyContactsLink() throws IOException {
		testUtil.switchToFrame();
		contactsPage = homePage.clickOnContactsLink();
		
	}

	@AfterMethod
	public void tearDown() {
		driver.close();
	}

}
